Style pass
Refactor SpriteSelector into "Target Pane" which contains the sprite selector, stage selector, and the buttons for new sprites, costumes and backdrops.
